Add files via upload
This is the project file set for extraction, transformation and loading of the data set. The two excel files are used for the ETL procedure and the code is split into two parts as project 1 and 2 - where project 1 pertains to the extraction part of the ETL procedure and project 2 pertains to the transformation and the loading part of the procedure.